Fix permission problems and improve docs
This can now be run via drake_installer, sudo drake_installer
was previously required when the installation directory
(/opt/drake) was not user owned, but drake_installer is not
disoverable in this mannger from the workspace install path.
It now switches out to sudo as needed.

This also improves the documentation, clearly delineating
cmake and python workflows.